Рефакторинг. Частичное документирование.

+++ Assets/Game/Scripts/ShipData.cs
@@ -3,8 +3,8 @@
 [Serializable]
 public class ShipData
 {
-    public int X { get; set; }
-    public int Y { get; set; }
+    public byte X { get; set; }
+    public byte Y { get; set; }
     public GameTileContentType Type;
     public Direction Direction { get; set; }
 }
